test(ci): split the session-protocol tests at the graph seam
The new test file reached 552 lines and tripped the taste file-size rule, taking
the tracked tree to 603 against a baseline of 602. The ratchet exists to block a
new error-severity violation, so raising the baseline was not the honest move.

The file already held two unrelated concerns. The script's own behaviour needs
only the tmp-dir fixture and a CompletedProcess helper. The workflow-graph
assertions need the AST closure walkers and nothing else. Splitting there leaves
358 and 216 lines and no shared helper crossing the seam.

51 tests pass across both files; taste-lints reports 0 errors on each; the count
ratchet test passes.
